+++ NEWS.md
@@ -1,5 +1,10 @@
 # Immigration Papers
 
+### Version 2.1.0 (2026-01-18)
+
+* Refactored all code to Python
+* use beta.paper-picnic.com as input
+
 ### Version 2.0.0 (2025-07-03)
 
 * Rank using finetuned SetFit model